EES-861 Moving NotificationService from Admin to the Publisher
