--- src/GovUk.Education.ExploreEducationStatistics.Admin/Services/Interfaces/INotificationsService.cs
@@ -1,9 +1,0 @@
-using System;
-
-namespace GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces
-{
-    public interface INotificationsService
-    {
-        bool NotifySubscribers(Guid publicationId);
-    }
-}
