+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Startup.cs
@@ -180,7 +180,6 @@ namespace GovUk.Education.ExploreEducationStatistics.Admin
 
             services.AddTransient<IFileStorageService, FileStorageService>();
             services.AddTransient<IImportService, ImportService>();
-            services.AddTransient<INotificationsService, NotificationsService>();
             services.AddTransient<IPublishingService, PublishingService>();
             services.AddTransient<IReleaseStatusService, ReleaseStatusService>();
             services.AddTransient<IThemeService, ThemeService>();
