+++ src/GovUk.Education.ExploreEducationStatistics.Publisher/Services/Interfaces/INotificationsService.cs
@@ -1,0 +1,10 @@
+using System;
+using System.Threading.Tasks;
+
+namespace GovUk.Education.ExploreEducationStatistics.Publisher.Services.Interfaces
+{
+    public interface INotificationsService
+    {
+        Task NotifySubscribersAsync(Guid publicationId);
+    }
+}
